Add files via upload
ONC Watch is written for digital health app developers.

This article gives some background about the Office of the National Coordinator for Health Information Technology (ONC), and why its activities matter to digital health app developers.  It also summarizes some key takeaways from the ONC's recently released Proposed Rulemaking for implementing provisions of the 21st Century Cures Act.  The ONC's proposed rule is 724-pages long (double spaced), and covers a lot of other proposals that might be interesting to some digital health app developers, and not to others.  This article is a way to get started with wrangling those 724-pages, with some orienting context and a few "pins".

This ONC Watch is written for digital health app developers.